# openapi/pufferlab-v1.json
{
  "components": {
    "schemas": {
      "ApiErrorCode": {
        "enum": [
          "validation_error",
          "not_found",
          "namespace_not_ready",
          "provider_error",
          "rate_limited",
          "run_conflict",
          "internal_error"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
      "ApiErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Output"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "trace_id"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
...
      "ContractVersion": {
        "const": 1,
        "type": "integer"
      },
...
      "DataOrigin": {
        "description": "Whether persisted catalog data came from provider-backed work or the offline demo.",
        "enum": [
          "live",
          "synthetic_demo"
        ],
        "title": "DataOrigin",
        "type": "string"
      },
...
      "DatasetCatalogItem": {
        "additionalProperties": false,
        "properties": {
          "data_origin": {
            "$ref": "#/components/schemas/DataOrigin"
          },
          "dataset": {
            "$ref": "#/components/schemas/DatasetVersion"
          }
        },
        "required": [
          "dataset",
          "data_origin"
        ],
        "title": "DatasetCatalogItem",
        "type": "object"
      },
      "DatasetDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "data_origin": {
            "$ref": "#/components/schemas/DataOrigin"
          },
          "dataset": {
            "$ref": "#/components/schemas/DatasetVersion"
          }
        },
        "required": [
          "dataset",
          "data_origin"
        ],
        "title": "DatasetDetailResponse",
        "type": "object"
      },
      "DatasetListResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetCatalogItem"
            },
            "maxItems": 100,
            "title": "Datasets",
            "type": "array"
          }
        },
        "required": [
          "datasets"
        ],
        "title": "DatasetListResponse",
        "type": "object"
      },
      "DatasetStatus": {
        "enum": [
          "pending",
          "ingesting",
          "indexing",
          "ready",
          "failed"
        ],
        "title": "DatasetStatus",
        "type": "string"
      },
      "DatasetVersion": {
        "additionalProperties": false,
        "properties": {
          "corpus_hash": {
            "minLength": 1,
            "title": "Corpus Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_origin": {
            "$ref": "#/components/schemas/DataOrigin",
            "default": "live"
          },
          "document_count": {
            "minimum": 0.0,
            "title": "Document Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "index_profile": {
            "$ref": "#/components/schemas/IndexProfile"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "slug": {
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DatasetStatus"
          },
          "version": {
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "version",
          "namespace",
          "index_profile",
          "document_count",
          "corpus_hash",
          "status",
          "created_at"
        ],
        "title": "DatasetVersion",
        "type": "object"
      },
...
      "FtsProfile": {
        "additionalProperties": false,
        "properties": {
          "ascii_folding": {
            "default": false,
            "title": "Ascii Folding",
            "type": "boolean"
          },
          "b": {
            "default": 0.75,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "B",
            "type": "number"
          },
          "case_sensitive": {
            "default": false,
            "title": "Case Sensitive",
            "type": "boolean"
          },
          "k1": {
            "default": 1.2,
            "exclusiveMinimum": 0.0,
            "title": "K1",
            "type": "number"
          },
          "k3": {
            "default": 8.0,
            "exclusiveMinimum": 0.0,
            "title": "K3",
            "type": "number"
          },
          "language": {
            "default": "english",
            "title": "Language",
            "type": "string"
          },
          "max_token_length": {
            "default": 39,
            "maximum": 255.0,
            "minimum": 1.0,
            "title": "Max Token Length",
            "type": "integer"
          },
          "remove_stopwords": {
            "default": false,
            "title": "Remove Stopwords",
            "type": "boolean"
          },
          "stemming": {
            "default": false,
            "title": "Stemming",
            "type": "boolean"
          },
          "tokenizer": {
            "default": "word_v4",
            "title": "Tokenizer",
            "type": "string"
          }
        },
        "title": "FtsProfile",
        "type": "object"
      },
...
      "IndexProfile": {
        "additionalProperties": false,
        "properties": {
          "distance_metric": {
            "enum": [
              "cosine_distance",
              "euclidean_squared"
            ],
            "title": "Distance Metric",
            "type": "string"
          },
          "embedding_model": {
            "minLength": 1,
            "title": "Embedding Model",
            "type": "string"
          },
          "embedding_provider": {
            "const": "sentence_transformers",
            "title": "Embedding Provider",
            "type": "string"
          },
          "embedding_revision": {
            "minLength": 1,
            "title": "Embedding Revision",
            "type": "string"
          },
          "fts_profile": {
            "$ref": "#/components/schemas/FtsProfile"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "schema_hash": {
            "minLength": 1,
            "title": "Schema Hash",
            "type": "string"
          },
          "vector_attribute": {
            "default": "vector",
            "title": "Vector Attribute",
            "type": "string"
          },
          "vector_dimensions": {
            "exclusiveMinimum": 0.0,
            "title": "Vector Dimensions",
            "type": "integer"
          },
          "vector_dtype": {
            "enum": [
              "f16",
              "f32",
              "i8"
            ],
            "title": "Vector Dtype",
            "type": "string"
          }
        },
        "required": [
          "id",
          "embedding_provider",
          "embedding_model",
          "embedding_revision",
          "vector_dimensions",
          "vector_dtype",
          "distance_metric",
          "fts_profile",
          "schema_hash"
        ],
        "title": "IndexProfile",
        "type": "object"
      },
...
      "JsonValue-Output": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue-Output"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Output"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "QuerySet": {
        "additionalProperties": false,
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_version_id": {
            "format": "uuid",
            "title": "Dataset Version Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "query_count": {
            "minimum": 0.0,
            "title": "Query Count",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "dataset_version_id",
          "query_count",
          "content_hash",
          "created_at"
        ],
        "title": "QuerySet",
        "type": "object"
      },
      "QuerySetCatalogItem": {
        "additionalProperties": false,
        "properties": {
          "data_origin": {
            "$ref": "#/components/schemas/DataOrigin"
          },
          "query_set": {
            "$ref": "#/components/schemas/QuerySet"
          }
        },
        "required": [
          "query_set",
          "data_origin"
        ],
        "title": "QuerySetCatalogItem",
        "type": "object"
      },
      "QuerySetListResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "dataset_version_id": {
            "format": "uuid",
            "title": "Dataset Version Id",
            "type": "string"
          },
          "query_sets": {
            "items": {
              "$ref": "#/components/schemas/QuerySetCatalogItem"
            },
            "maxItems": 100,
            "title": "Query Sets",
            "type": "array"
          }
        },
        "required": [
          "dataset_version_id",
          "query_sets"
        ],
        "title": "QuerySetListResponse",
        "type": "object"
      },
...
      "RetrievalConfigCatalogResponse": {
        "additionalProperties": false,
        "properties": {
          "configs": {
            "items": {
              "$ref": "#/components/schemas/RetrievalConfigSummary"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Configs",
            "type": "array"
          },
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "data_origin": {
            "$ref": "#/components/schemas/DataOrigin"
          },
          "dataset_version_id": {
            "format": "uuid",
            "title": "Dataset Version Id",
            "type": "string"
          }
        },
        "required": [
          "dataset_version_id",
          "data_origin",
          "configs"
        ],
        "title": "RetrievalConfigCatalogResponse",
        "type": "object"
      },
...
      "RetrievalConfigSummary": {
        "additionalProperties": false,
        "properties": {
          "config_hash": {
            "title": "Config Hash",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/RetrievalMode"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "revision",
          "name",
          "mode",
          "config_hash"
        ],
        "title": "RetrievalConfigSummary",
        "type": "object"
      },
      "RetrievalMode": {
        "enum": [
          "bm25",
          "vector",
          "hybrid_rrf",
          "hybrid_rerank"
        ],
        "title": "RetrievalMode",
        "type": "string"
      },
...
    }
  },
  "info": {
    "summary": "Search evaluation and query-forensics workbench",
    "title": "PufferLab API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datasets": {
      "get": {
        "operationId": "list_evaluation_datasets",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested revision was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The persisted catalog is not canonical."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The catalog request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Stored catalog data is unavailable."
          }
        },
        "summary": "List Evaluation Datasets",
        "tags": [
          "evaluation catalogs"
        ]
      }
    },
    "/api/v1/datasets/{dataset_version_id}": {
      "get": {
        "operationId": "get_evaluation_dataset",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested revision was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The persisted catalog is not canonical."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The catalog request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Stored catalog data is unavailable."
          }
        },
        "summary": "Get Evaluation Dataset",
        "tags": [
          "evaluation catalogs"
        ]
      }
    },
    "/api/v1/datasets/{dataset_version_id}/configs": {
      "get": {
        "operationId": "list_dataset_evaluation_configs",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalConfigCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested revision was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The persisted catalog is not canonical."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The catalog request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Stored catalog data is unavailable."
          }
        },
        "summary": "List Dataset Evaluation Configs",
        "tags": [
          "evaluation catalogs"
        ]
      }
    },
...
    "/api/v1/query-sets": {
      "get": {
        "operationId": "list_evaluation_query_sets",
        "parameters": [
          {
            "in": "query",
            "name": "dataset_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuerySetListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested revision was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The persisted catalog is not canonical."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The catalog request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Stored catalog data is unavailable."
          }
        },
        "summary": "List Evaluation Query Sets",
        "tags": [
          "evaluation catalogs"
        ]
      }
    },
...
  }
}